refactor: change project structure
Change structure from Model-View-Controller to a less abstract
DBMS-Server model
