+++ src/config.h
@@ -6,7 +6,7 @@
 #include <variant>
 #include <glog/logging.h>
 
-#include "Model/Cell/Cell.h"
+#include "DBMS/Cell/Cell.h"
 
 static const std::string PROJECT_ROOT (std::getenv("SCRATCHQL_ROOT"));
