+++ test/DBMS/FileHandler/FileHandler.cpp
@@ -1,5 +1,5 @@
 #include "../../main.h"
-#include "../../../src/Model/FileHandler/FileHandler.h"
+#include "../../../src/DBMS/FileHandler/FileHandler.h"
 
 SCENARIO ("Reading from and writing to a file is executed correctly") {
     GIVEN ("A file and some test strings") {
